Common arguments of laboratory processing functions

Common arguments of laboratory processing functions

Arguments

labData
a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName
the column name that records patient ID in labData.

labItemColName
the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

dateColName
the column name that records test date in labData. It should be in "YYYYMMDD"/"YYYY-MM-DD" format.

indexDate
the specific date that used for cutting time window. It can be first record ("First"), last record ("Last"), or any date of interest with "YYYYMMDD"/"YYYY-MM-DD" format.

gapDate
desired period (in days) of each window interval. If NULL, it will be seen as only one single time window.

completeWindows
logical. If TRUE, time window series will be complete in order. If TRUE, a window without lab test will be skipped. Default is TRUE.