Obtaining Time-Series Data

getTimeSeriesLab seperates original data into multiple time windows, and summarize statistical information.

getTimeSeriesLab(
  labData,
  idColName,
  labItemColName,
  dateColName,
  valueColName,
  indexDate = last,
  gapDate = NULL,
  completeWindows = TRUE
)

Arguments

labData
a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName
the column name that records patient ID in labData.

labItemColName
the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

dateColName
the column name that records test date in labData. It should be in "YYYYMMDD"/"YYYY-MM-DD" format.

valueColName
the column name that records test value in labData. Only numeric value is acceptable.

indexDate
the specific date that used for cutting time window. It can be first record ("first"), last record ("last"), any date of interest with "YYYYMMDD"/"YYYY-MM-DD" format, or any self-defined mapping table. When using self-defined table, make sure that thee table contains two column with column name: 'ID' and 'indexDate'.

gapDate
desired period (in days) of each window interval. If NULL, it will be seen as only one single time window.

completeWindows
logical. If TRUE, time window series will be complete in order. If TRUE, a window without lab test will be skipped. Default is TRUE.

Value

A data.table with statistical summary.

Details

This function is used for seperating lab data into multiple time windows, and it provides overall statistical information: total count, maximun value, minimun value, mean, nearest record to index date of each time window. If indexDate is first, then it will be the earliest test date among all the lab tests.

Examples

# set last record as index date
getTimeSeriesLab(labData = labSample,
                idColName = SUBJECT_ID,
                labItemColName = ITEMID,
                dateColName = CHARTTIME,
                valueColName = VALUENUM,
                indexDate = last,
                gapDate = 360,
                completeWindows = TRUE)

#>      ID ITEMID Window Count    Max   Min        Mean Nearest firstRecord
#>  1:  36  50811     -4     4   12.7  10.0  11.6750000    12.7  2131-05-04
#>  2:  36  50811     -3    NA     NA    NA          NA      NA        <NA>
#>  3:  36  50811     -2    NA     NA    NA          NA      NA        <NA>
#>  4:  36  50811     -1    NA     NA    NA          NA      NA        <NA>
#>  5:  36  50861     -4     2   12.0   8.0  10.0000000    12.0  2131-04-30
#>  6:  36  50861     -3    NA     NA    NA          NA      NA        <NA>
#>  7:  36  50861     -2    NA     NA    NA          NA      NA        <NA>
#>  8:  36  50861     -1     1   12.0  12.0  12.0000000    12.0  2134-05-14
#>  9:  36  50912     -4    23    2.0   0.7   1.2347826     1.3  2131-04-30
#> 10:  36  50912     -3     5    1.2   1.1   1.1400000     1.1  2131-06-22
#> 11:  36  50912     -2    NA     NA    NA          NA      NA        <NA>
#> 12:  36  50912     -1     9    1.0   0.8   0.8888889     1.0  2134-05-11
#> 13:  36  50931     -4    19  161.0  82.0 107.5263158   103.0  2131-04-30
#> 14:  36  50931     -3     4  116.0  75.0  91.5000000    75.0  2131-06-22
#> 15:  36  50931     -2    NA     NA    NA          NA      NA        <NA>
#> 16:  36  50931     -1     9  126.0  85.0 111.3333333    85.0  2134-05-11
#> 17:  36  51222     -4    22   14.6   7.1  10.2090909     8.8  2131-04-30
#> 18:  36  51222     -3     5   14.5  11.3  12.9200000    12.9  2131-06-22
#> 19:  36  51222     -2    NA     NA    NA          NA      NA        <NA>
#> 20:  36  51222     -1     8   14.4  10.9  11.9375000    11.8  2134-05-11
#> 21: 109  50811     -5    NA     NA    NA          NA      NA        <NA>
#> 22: 109  50811     -4    NA     NA    NA          NA      NA        <NA>
#> 23: 109  50811     -3    NA     NA    NA          NA      NA        <NA>
#> 24: 109  50811     -2    NA     NA    NA          NA      NA        <NA>
#> 25: 109  50811     -1     2    7.2   6.9   7.0500000     6.9  2141-09-20
#> 26: 109  50861     -5     4   51.0  13.0  26.2500000    14.0  2137-12-04
#> 27: 109  50861     -4    NA     NA    NA          NA      NA        <NA>
#> 28: 109  50861     -3     3   38.0  20.0  28.0000000    20.0  2140-04-29
#> 29: 109  50861     -2     4   80.0  14.0  31.5000000    80.0  2141-03-27
#> 30: 109  50861     -1    29   78.0   8.0  20.6551724    19.0  2141-09-20
#> 31: 109  50912     -5    33   13.6   3.9   6.0060606     6.0  2137-11-04
#> 32: 109  50912     -4     2    6.9   6.3   6.6000000     6.9  2138-10-03
#> 33: 109  50912     -3    38    6.6   2.6   4.4263158     5.2  2140-01-19
#> 34: 109  50912     -2    23    8.3   6.2   7.5217391     7.1  2141-03-27
#> 35: 109  50912     -1   142   10.5   3.5   6.4478873     5.7  2141-09-05
#> 36: 109  50931     -5    33  151.0  64.0 100.0909091    97.0  2137-11-04
#> 37: 109  50931     -4     2  106.0 105.0 105.5000000   105.0  2138-10-03
#> 38: 109  50931     -3    38  150.0  70.0 103.1315789    79.0  2140-01-19
#> 39: 109  50931     -2    22  256.0  70.0 101.6363636    90.0  2141-03-28
#> 40: 109  50931     -1   140  154.0  67.0  93.3000000    91.0  2141-09-05
#> 41: 109  51222     -5    35   15.7   5.7   9.5114286     9.7  2137-11-04
#> 42: 109  51222     -4    NA     NA    NA          NA      NA        <NA>
#> 43: 109  51222     -3    39   14.4   8.0  10.6076923     9.4  2140-01-19
#> 44: 109  51222     -2    23    8.8   6.0   7.0260870     7.1  2141-03-27
#> 45: 109  51222     -1   138   13.1   6.0   8.0862319     6.8  2141-09-05
#> 46: 132  50861     -1     2   59.0  22.0  40.5000000    59.0  2115-05-06
#> 47: 132  50912     -1    32    0.6   0.2   0.4062500     0.5  2115-05-06
#> 48: 132  50931     -1    32  138.0  70.0  91.7500000    77.0  2115-05-06
#> 49: 132  51222     -1    26   15.1   9.0  11.9076923    13.3  2115-05-06
#> 50: 143  50811     -1     1   12.7  12.7  12.7000000    12.7  2155-01-18
#> 51: 143  50861     -1    60 1226.0  15.0 250.1166667    31.0  2154-12-25
#> 52: 143  50904     -1     1   51.0  51.0  51.0000000    51.0  2155-12-10
#> 53: 143  50906     -1     1  120.0 120.0 120.0000000   120.0  2155-12-10
#> 54: 143  50912     -1    60    2.1   0.6   1.0716667     1.2  2154-12-25
#> 55: 143  50931     -1    57  432.0  76.0 119.8596491    93.0  2154-12-25
#> 56: 143  51222     -1    62   13.9   7.6  10.7241935    10.4  2154-12-25
#> 57: 145  50811     -1     9   13.0   7.0  10.0555556    11.2  2144-03-30
#> 58: 145  50861     -1    74  252.0  14.0  81.1216216    60.0  2144-03-29
#> 59: 145  50904     -1     2   32.0  20.0  26.0000000    20.0  2144-04-07
#> 60: 145  50906     -1     1  108.0 108.0 108.0000000   108.0  2144-04-07
#> 61: 145  50912     -1   162   10.8   0.4   1.5537037     1.2  2144-03-29
#> 62: 145  50931     -1   159  231.0  65.0 104.8176101   111.0  2144-03-29
#> 63: 145  51222     -1   142   12.9   7.0   9.7802817    10.1  2144-03-29
#>      ID ITEMID Window Count    Max   Min        Mean Nearest firstRecord
#>     lastRecode
#>  1: 2131-05-18
#>  2:       <NA>
#>  3:       <NA>
#>  4:       <NA>
#>  5: 2131-05-17
#>  6:       <NA>
#>  7:       <NA>
#>  8: 2134-05-14
#>  9: 2131-05-25
#> 10: 2132-02-08
#> 11:       <NA>
#> 12: 2134-05-20
#> 13: 2131-05-25
#> 14: 2132-02-08
#> 15:       <NA>
#> 16: 2134-05-20
#> 17: 2131-05-25
#> 18: 2132-02-08
#> 19:       <NA>
#> 20: 2134-05-18
#> 21:       <NA>
#> 22:       <NA>
#> 23:       <NA>
#> 24:       <NA>
#> 25: 2142-02-25
#> 26: 2138-07-03
#> 27:       <NA>
#> 28: 2140-05-17
#> 29: 2141-06-11
#> 30: 2142-07-22
#> 31: 2138-08-19
#> 32: 2138-10-27
#> 33: 2140-09-02
#> 34: 2141-06-17
#> 35: 2142-08-30
#> 36: 2138-08-19
#> 37: 2138-10-27
#> 38: 2140-09-02
#> 39: 2141-06-17
#> 40: 2142-08-30
#> 41: 2138-08-20
#> 42:       <NA>
#> 43: 2140-09-02
#> 44: 2141-06-17
#> 45: 2142-08-30
#> 46: 2115-10-04
#> 47: 2116-04-08
#> 48: 2116-04-08
#> 49: 2116-04-08
#> 50: 2155-01-18
#> 51: 2155-10-22
#> 52: 2155-12-10
#> 53: 2155-12-10
#> 54: 2155-10-22
#> 55: 2155-10-07
#> 56: 2155-12-10
#> 57: 2145-02-15
#> 58: 2145-02-16
#> 59: 2144-05-28
#> 60: 2144-04-07
#> 61: 2145-02-22
#> 62: 2145-02-22
#> 63: 2145-02-18
#>     lastRecode

# use index date mapping table

indexDateTable <- data.table(ID = c(36, 109, 145, 143, 132),
                        indexDate = c("2131-05-16", "2141-09-20", "2144-03-31", "2155-04-05", "2115-06-01"))
getTimeSeriesLab(labData = loincSample,
                idColName = SUBJECT_ID,
                labItemColName = LOINC + CATEGORY,
                dateColName = CHARTTIME,
                valueColName = VALUENUM,
                indexDate = indexDateTable,
                gapDate = 30,
                completeWindows = TRUE)

#>       ID  LOINC   CATEGORY Window Count  Max  Min      Mean Nearest firstRecord
#>   1:  36 1742-6  Chemistry     -1     1  8.0  8.0  8.000000     8.0  2131-04-30
#>   2:  36 1742-6  Chemistry      1     1 12.0 12.0 12.000000    12.0  2131-05-17
#>   3:  36 1742-6  Chemistry      2    NA   NA   NA        NA      NA        <NA>
#>   4:  36 1742-6  Chemistry      3    NA   NA   NA        NA      NA        <NA>
#>   5:  36 1742-6  Chemistry      4    NA   NA   NA        NA      NA        <NA>
#>  ---                                                                           
#> 709: 145  718-7 Hematology      7     9 12.0  9.1 10.600000     9.1  2144-09-27
#> 710: 145  718-7 Hematology      8    NA   NA   NA        NA      NA        <NA>
#> 711: 145  718-7 Hematology      9    14 10.7  8.9  9.528571    10.1  2144-12-06
#> 712: 145  718-7 Hematology     10    18 12.9 10.3 11.333333    12.9  2144-12-28
#> 713: 145  718-7 Hematology     11     4 12.0  9.8 10.700000    12.0  2145-02-15
#>      lastRecode
#>   1: 2131-04-30
#>   2: 2131-05-17
#>   3:       <NA>
#>   4:       <NA>
#>   5:       <NA>
#>  ---           
#> 709: 2144-10-05
#> 710:       <NA>
#> 711: 2144-12-23
#> 712: 2145-01-14
#> 713: 2145-02-18