Mapping local lab codes with LOINC

mapLOINC is used for merging two tables together. One is lab data with local codes, and the other one is mapping table of local lab codes and LOINC.

mapLOINC(labData, labItemColName, mappingTable = NULL)

Arguments

labData
a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

labItemColName
the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

mappingTable
a data table with at least two columns: local lab item code and LOINC code. The local column name should be identical as labItemColName in labData.

Value

A data.table merged two tables.

Details

Since the lab item may be coded in local schemes, users can map it with LOINC by providing LOINC mapping table. This step is optional but recommended. After mapped, user can select desired laboratory tests by using function searchCases.

Examples

mapLOINC(labData = labMIMIC,
        labItemColName = ITEMID,
        mappingTable = mapMIMIC)

#>         ITEMID ROW_ID.x SUBJECT_ID HADM_ID           CHARTTIME VALUE VALUENUM
#>      1:  50800      282          3      NA 2101-10-12 18:17:00   ART       NA
#>      2:  50800      394          3      NA 2101-10-15 12:42:00   ART       NA
#>      3:  50800      966          3  145834 2101-10-22 04:31:00   ART       NA
#>      4:  50800      984          3  145834 2101-10-22 07:13:00   ART       NA
#>      5:  50800      994          3  145834 2101-10-22 10:16:00   ART       NA
#>     ---                                                                      
#>  99996:  51523    90310        147  103631 2158-06-24 17:20:00  HOLD       NA
#>  99997:  51523    93500        151  151029 2145-05-06 19:39:00  HOLD       NA
#>  99998:  51523    92888        149  154869 2135-02-18 10:30:00  HOLD       NA
#>  99999:  51523    95640        155  147148 2125-09-10 10:45:00  HOLD       NA
#> 100000:  51529    24741         55      NA 2140-03-22 08:00:00   126      126
#>         VALUEUOM     FLAG ROW_ID.y                    LABEL FLUID  CATEGORY
#>      1:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      2:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      3:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      4:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      5:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>     ---                                                                    
#>  99996:                        723                  GR HOLD URINE CHEMISTRY
#>  99997:                        723                  GR HOLD URINE CHEMISTRY
#>  99998:                        723                  GR HOLD URINE CHEMISTRY
#>  99999:                        723                  GR HOLD URINE CHEMISTRY
#> 100000:    mg/dL abnormal      729 Estimated Actual Glucose BLOOD CHEMISTRY
#>         LOINC_CODE
#>      1:           
#>      2:           
#>      3:           
#>      4:           
#>      5:           
#>     ---           
#>  99996:           
#>  99997:           
#>  99998:           
#>  99999:           
#> 100000: