Wide Time-Series Data

wideTimeSeriesLab transforms time-series data into wide format.

wideTimeSeriesLab(
  labData,
  idColName,
  labItemColName,
  windowColName,
  valueColName
)

Arguments

labData

a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName

the column name that records patient ID in labData.

labItemColName

the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

windowColName

the column name that records time window sequence in labData.

Value

A new, wide-formatted data.table.

Details

After setting desired time window, data can further transform into wide format, which is more suitable for analysis. If lab item code is composed by multiple columns, it will automaticlly combined as one. And there can only be single value column selected in this step.

Examples

timeSeriesData <- getTimeSeriesLab(labData = labSample,
                                   idColName = SUBJECT_ID,
                                   labItemColName = ITEMID,
                                   dateColName = CHARTTIME,
                                   valueColName = VALUENUM,
                                   indexDate = first,
                                   gapDate = 360,
                                   completeWindows = TRUE)
imputedData <- imputeTimeSeriesLab(labData = timeSeriesData,
                                   idColName = ID,
                                   labItemColName = ITEMID,
                                   windowColName = Window,
                                   valueColName = Max & Min,
                                   impMethod = mean,
                                   imputeOverallMean=FALSE)
#> Error in set(x, j = name, value = value): RHS of assignment to existing column 'imputed' is zero length but not NULL. If you intend to delete the column use NULL. Otherwise, the RHS must have length > 0; e.g., NA_integer_. If you are trying to change the column type to be an empty list column then, as with all column type changes, provide a full length RHS vector such as vector('list',nrow(DT)); i.e., 'plonk' in the new column.
wideTimeSeriesLab(labData = imputedData,
                  idColName = ID,
                  labItemColName = ITEMID,
                  windowColName = Window,
                  valueColName = Max)
#> Error in as.data.table(labData): object 'imputedData' not found