Get abnormal mark

getAbnormalMark generates a new column recorded abnormal values: high H and L.

getAbnormalMark(
  labData,
  idColName,
  labItemColName,
  valueColName,
  genderColName,
  genderTable,
  referenceTable = refLOINC
)

Arguments

labData
a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName
the column name that records patient ID in labData.

labItemColName
the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

valueColName
the column name that records test value in labData. Only numeric value is acceptable.

genderColName
the column name that records gender of each patient. Gender should be coded as M or F.

referenceTable
a data table that records reference values of each lab item. Column names should be: LowerBound_Male, UpperBound_Male, LowerBound_Female, UpperBound_Female.

Value

A data.table merged two tables.

Details

This step is optional but recommended. Reference ranges differ in different institutions, so a user-defined reference table is needed. The table should record the reference values for male and female separately. Lower bound can be expressed as <, ( (greater than) or <= , [ (great tham or euqal to), and the expression of upper bound is similar. If there is no upper bound or lower bound, just remain empty. After the new abnormal flag column generated, it can be used to do categorical analysis, or be used in the function: plotTimeSeriesLab.

Examples

mapLOINC(labData = labMIMIC,
        labItemColName = ITEMID,
        mappingTable = mapMIMIC)

#>         ITEMID ROW_ID.x SUBJECT_ID HADM_ID           CHARTTIME VALUE VALUENUM
#>      1:  50800      282          3      NA 2101-10-12 18:17:00   ART       NA
#>      2:  50800      394          3      NA 2101-10-15 12:42:00   ART       NA
#>      3:  50800      966          3  145834 2101-10-22 04:31:00   ART       NA
#>      4:  50800      984          3  145834 2101-10-22 07:13:00   ART       NA
#>      5:  50800      994          3  145834 2101-10-22 10:16:00   ART       NA
#>     ---                                                                      
#>  99996:  51523    90310        147  103631 2158-06-24 17:20:00  HOLD       NA
#>  99997:  51523    93500        151  151029 2145-05-06 19:39:00  HOLD       NA
#>  99998:  51523    92888        149  154869 2135-02-18 10:30:00  HOLD       NA
#>  99999:  51523    95640        155  147148 2125-09-10 10:45:00  HOLD       NA
#> 100000:  51529    24741         55      NA 2140-03-22 08:00:00   126      126
#>         VALUEUOM     FLAG ROW_ID.y                    LABEL FLUID  CATEGORY
#>      1:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      2:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      3:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      4:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>      5:                          1            SPECIMEN TYPE BLOOD BLOOD GAS
#>     ---                                                                    
#>  99996:                        723                  GR HOLD URINE CHEMISTRY
#>  99997:                        723                  GR HOLD URINE CHEMISTRY
#>  99998:                        723                  GR HOLD URINE CHEMISTRY
#>  99999:                        723                  GR HOLD URINE CHEMISTRY
#> 100000:    mg/dL abnormal      729 Estimated Actual Glucose BLOOD CHEMISTRY
#>         LOINC_CODE
#>      1:           
#>      2:           
#>      3:           
#>      4:           
#>      5:           
#>     ---           
#>  99996:           
#>  99997:           
#>  99998:           
#>  99999:           
#> 100000: