Obtaining Time-Series Data

getTimeSeriesLab seperates original data into multiple time windows, and summarize statistical information.

getTimeSeriesLab(labData, idColName, labItemColName, dateColName,
  valueColName, indexDate = last, gapDate = NULL,
  completeWindows = TRUE)

Arguments

labData
a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName
the column name that records patient ID in labData.

labItemColName
the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add + operator between column names, e.g., A + B.

dateColName
the column name that records test date in labData. It should be in "YYYYMMDD"/"YYYY-MM-DD" format.

valueColName
the column name that records test value in labData. Only numeric value is acceptable.

indexDate
the specific date that used for cutting time window. It can be first record ("First"), last record ("Last"), or any date of interest with "YYYYMMDD"/"YYYY-MM-DD" format.

gapDate
desired period (in days) of each window interval. If NULL, it will be seen as only one single time window.

completeWindows
logical. If TRUE, time window series will be complete in order. If TRUE, a window without lab test will be skipped. Default is TRUE.

Value

A data.table with statistical summary.

Details

This function is used for seperating lab data into multiple time windows, and it provides overall statistical information: total count, maximun value, minimun value, mean, nearest record to index date of each time window.

Examples

getTimeSeriesLab(labData = labMIMIC,
                idColName = SUBJECT_ID,
                labItemColName = ITEMID,
                dateColName = CHARTTIME,
                valueColName = VALUENUM,
                indexDate = first,
                gapDate = 360,
                completeWindows = TRUE)

#>           ID ITEMID Window Count  Max  Min Mean Nearest firstRecord lastRecode
#>     1:     2  50883      1     1  0.3  0.3  0.3     0.3  2138-07-20 2138-07-20
#>     2:     2  50884      1     1  9.0  9.0  9.0     9.0  2138-07-20 2138-07-20
#>     3:     2  50885      1     1  9.3  9.3  9.3     9.3  2138-07-20 2138-07-20
#>     4:     2  51137      1     1   NA   NA   NA      NA  2138-07-17 2138-07-17
#>     5:     2  51143      1     2  0.0  0.0  0.0     0.0  2138-07-17 2138-07-17
#>    ---                                                                        
#> 20260:   171  51508      1     2   NA   NA   NA      NA  2197-07-20 2197-07-21
#> 20261:   171  51514      1     2   NA   NA   NA      NA  2197-07-20 2197-07-21
#> 20262:   171  51516      1     2  3.0  3.0  3.0     3.0  2197-07-20 2197-07-21
#> 20263:   171  51519      1     2   NA   NA   NA      NA  2197-07-20 2197-07-21
#> 20264: 99096  51265      1     1 32.0 32.0 32.0    32.0  2110-07-07 2110-07-07