KNN Imputation

imputeKNN uses k-nearest neighbors (kNN) method to fill missing values.

imputeKNN(labData, idColName, k)

Arguments

labData

a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.

idColName

the column name that records patient ID in labData.

k

the numbers of selected nearestZneighbors.

Value

A new, wide-formatted imputed data.table.

Details

Given wide format data, data can conduct KNN imputation. If lab item code are composed by multiple columns, it will automatically combined as one. All other columns will all be seen as value columns.

Examples

timeSeriesData <- getTimeSeriesLab(labData = labSample,
                                   idColName = SUBJECT_ID,
                                   labItemColName = ITEMID,
                                   dateColName = CHARTTIME,
                                   valueColName = VALUENUM,
                                   indexDate = first,
                                   gapDate = 360,
                                   completeWindows = TRUE)
imputedData <- imputeTimeSeriesLab(labData = timeSeriesData,
                                   idColName = ID,
                                   labItemColName = ITEMID,
                                   windowColName = Window,
                                   valueColName = Max & Min,
                                   impMethod = mean,
                                   imputeOverallMean=FALSE)
wideData <- wideTimeSeriesLab(labData = imputedData,
                  idColName = ID,
                  labItemColName = ITEMID,
                  windowColName = Window,
                  valueColName = Max)

imputeKNN(labData = wideData,
          idColName = ID + Window,
          k = 2)
#>      ID Window 50811   50861    50904    50906  50912  50931  51222
#>  1:  36      1 12.70   12.00 41.50000 114.0000  2.000 161.00 14.600
#>  2:  36      2 12.70   12.00 36.75000 111.0000  1.500 143.50 14.500
#>  3:  36      3 12.70   12.00 39.12500 112.5000  1.500 143.50 14.500
#>  4:  36      4 12.70   12.00 37.93750 111.7500  1.000 126.00 14.400
#>  5: 109      1  7.05   51.00 37.93750 111.7500 13.600 151.00 15.700
#>  6: 109      2  7.05   61.75 39.71875 112.8750  9.875 177.75 13.625
#>  7: 109      3  7.05   38.00 38.82812 112.3125  6.600 150.00 14.400
#>  8: 109      4  7.20   80.00 38.82812 112.3125  8.800 256.00 11.300
#>  9: 109      5  6.90   78.00 38.82812 112.3125 10.500 154.00 13.100
#> 10: 132      1  7.05   59.00 38.38281 112.0312  0.600 138.00 15.100
#> 11: 143      1 12.70 1226.00 51.00000 120.0000  2.100 432.00 13.900
#> 12: 145      1 13.00  252.00 32.00000 108.0000 10.800 231.00 12.900